+++ tests/pipelines/test_tyre_impression_processing_pipeline.py
@@ -1,8 +1,9 @@
 import os
 import pytest
-from unittest.mock import MagicMock
-from domain.exceptions import PipelineError, ProcessorError
+import numpy as np
+from unittest.mock import MagicMock, patch
 from database.models.data_types.files import FileModel
+from domain.exceptions import PipelineError, ProcessorError
 from tests.helpers.factories.file_factory import FileFactory
 from tests.helpers.factories.tyre_impression_factory import TyreImpressionFactory
 from pipelines.tyre_impression_processing_pipeline import TyreImpressionProcessingPipeline
@@ -36,11 +37,23 @@ def test_tyre_impression_processing_pipeline_process_invalid_original_file_locat
         file_location="/invalid/file/location"
     )
 
-    with pytest.raises(
-            PipelineError,
-            match="TyreImpressionProcessing original image location does not exist"
-    ):
+    with pytest.raises(PipelineError, match="TyreImpressionProcessing original image location does not exist"):
         pipeline.process(tyre_impression_processing.id)
+
+
+def test_process_invalid_file_read_from_cv2_imread(pipeline):
+    tyre_impression = TyreImpressionFactory.create()
+    tyre_impression_processing = TyreImpressionProcessingFactory.create(tyre_impression.id)
+    os.makedirs("/files/test_directory", exist_ok=True)
+    FileFactory.create(
+        model=FileModel.tyre_impression,
+        model_id=tyre_impression_processing.id,
+        file_location="/files/test_directory"
+    )
+
+    with patch("cv2.imread", return_value=None):
+        with pytest.raises(PipelineError, match="Unable to read image in tyre impression processing pipeline"):
+            pipeline.process(tyre_impression_processing.id)
 
 
 def test_tyre_impression_processing_pipeline_processor_error(pipeline):
@@ -64,20 +77,20 @@ def test_tyre_impression_processing_pipeline_processor_error(pipeline):
         file_location="/files/test_directory"
     )
 
-    with pytest.raises(PipelineError, match="ProcessorError from stage1 processor: test error"):
-        result = pipeline.process(tyre_impression_processing.id)
-        assert result is None
+    with patch("cv2.imread", return_value=np.zeros((10,10))):
+        with pytest.raises(PipelineError, match="ProcessorError from stage1 processor: test error"):
+            pipeline.process(tyre_impression_processing.id)
 
 
 def test_tyre_impression_processing_pipeline_process(pipeline):
     # Mock pipeline processor stages
     stage1 = MagicMock()
     stage1.name = "stage1"
-    stage1.process.return_value = "path1"
+    stage1.process.return_value = np.ones((10,10))
 
     stage2 = MagicMock()
     stage2.name = "stage2"
-    stage2.process.return_value = "path2"
+    stage2.process.return_value = np.random.randint((10,10))
 
     pipeline.stages = [stage1, stage2]
 
@@ -90,12 +103,13 @@ def test_tyre_impression_processing_pipeline_process(pipeline):
         file_location="/files/test_directory"
     )
 
-    result = pipeline.process(tyre_impression_processing.id)
+    with patch("cv2.imread", return_value=np.zeros((10,10))):
+        result = pipeline.process(tyre_impression_processing.id)
 
-    assert stage1.process.called
-    assert stage2.process.called
+    stage1.process.assert_called_once()
+    assert np.array_equal(stage1.process.call_args[0][0], np.zeros((10,10)))
 
-    assert stage1.process.call_args[0][0] == original_file.file_location
-    assert stage2.process.call_args[0][0] == "path1"
+    stage2.process.assert_called_once()
+    assert np.array_equal(stage2.process.call_args[0][0], stage1.process.return_value)
 
-    assert "path2" == result
+    assert np.array_equal(stage2.process.return_value, result)